DDF-1964: Fixed Source UI bugs dealing with federating and publishing
Fixed bug when saving values on first add of the first configuration added.
Fixed scoping problem in Service.js when not publishing node to a registry
Removed activeBindingSelect from soureModal. This selection was redundant since it is shown on the main source page.
Removed Node type from source row. This will be put in in a later iteration.
Fixed bug with registry checkboxes not displaying for a connected registry when that connected registry contains no nodes.
Fixed bug in ModalSource.view.js in which disabled services were having their fpid changed as if they were becoming active configurations.