DDF-2267 Registry UI to Display All Nodes
Changing FederationAdmin's deleteLocalEntry to call getRegistryMetacardsByIds, updating unit tests

Adding check to see if the construct title is a string or a function
Adding check to see if the autoPopulateFunction  is a string or a function

Changing registry node model url to call allRegistryMetacards instead of localNodes
Adding getRemoteNodes function to get the non local entries from the model

Adding a remoteNodeRegion that will display the non local nodes.
Adding a function to display the NodeModel in readOnly mode
Adding a readOnly flag that is set when NodeTable is created and passed to the necessary views. The readOnly flag is used to prevent modifications to the non local entries.

Changing addSegment in the view to stop the event propagation
Changing the css to add the remoteNodeRegion and line things up

Changing the handlebars files to only allow fields to be editable and display add/remove icons if the readOnly flag is not set
(cherry picked from commit fb0ad74)